misisng cast
git-svn-id: https://svn.r-project.org/R/trunk@64408 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/stats/src/lm.c
@@ -59,7 +59,7 @@ SEXP Cdqrls(SEXP x, SEXP y, SEXP tol, SEXP chk)
     n = dims[0];
     p = dims[1];
     if (n)
-        ny = XLENGTH(y) / n; /* y :  n x ny, or an n - vector */
+        ny = (int)(XLENGTH(y) / n); /* y :  n x ny, or an n - vector */
     if (check && n * ny != XLENGTH(y))
         error(_("dimensions of 'x' (%d,%d) and 'y' (%d) do not match"), n, p, XLENGTH(y));
